W18 fee sweep
Push new json files for this week, and change address of the collector in the template as it is 0xef on sidechains  and 0x5e on mainnet, but mainnet is never handled via multisig sweep.

This probs needs more refactoring but is likely not to get used much soon due to mimic.