BIP-XXX Add Child Chain Gauge Factory to Psuedominter on AVAX
The Dao multisig must grant and revoke the addGaugeFactory role using the Authorizer. This permits the child chain gauge factory to be added to the psuedominter. This is needed in order for users to be able to claim their minted BAL from the gauges on Avax.

Simulation: https://dashboard.tenderly.co/public/safe/safe-apps/simulator/b292c605-81c1-4e3b-8639-30a2aca4235d