USDC_Arbi_CowPools_Extension
Owner: Zen Dragon

Take time reviewing as this is a fund handling payload. It will not be loaded until approved and merged. Goal is to bridge 20,000 USDC from the LM Multisig on mainnet to the LM Multisig on Arbitrum. This will be distributed at discretion of the Maxis towards Cow AMM pools in exchange for BAL which will be distributed in it's place towards veBAL holders.