sdBAL-Mint-and-Stake
Transaction uses all 80BAL-20wETH BPT from payroll to approve stake dao BalancerDepositor as a spender, then depositAll is called on the contract with lock and stake set to true to earn excess rewards. Once our position is created we have the option to delegate our position and boost if it is sensible.