auraBAL-Bridge-to-Arbi
For Q4 2024 to Q1 2025 we will be able to seed the vesters for auraBAL opt in payees. Next step will be to stake auraBAL on Arbitrum, break down amounts proportionally and begin seeding vesters for respective months and time.

Revoke of apporvals of BAL8020 to sdBAL and BAL to Cowswap are included in this bundle for house keeping.

Will not be loaded until approved.