Send-Funds-to-Injectors-v3-Launch
Final step to send rewards to all planned incentives pools. More which are configured to be added to schedule at a later date via injectors. Due to trust of v1 injector we send all incentives to respective addresses on mainnet. For v2 injector on gnosis only 2 weeks worth of funds for now.